Main

Description

A 2-player Tic Tac Toe game.

Documentation

main :: IO ()

Print a title, ask for the players' names, print some instructions and start the game.

printInstructions :: IO ()

Print the game instructions.

playRound :: PlayerName -> PlayerName -> Game -> IO ()

Execute a round of the game.

takeTurn :: Mark -> PlayerName -> Game -> IO Game

Prompt X or O for a move, returning an updated game.

drawGame :: Game -> IO ()

Draw a game on the screen.

describeMove :: PlayerName -> Cell -> IO ()

Print some information about a player's move at a cell.

gameOver :: PlayerName -> PlayerName -> Game -> IO ()

Draw the final game and congratulate the winner.

promptForName :: String -> IO String

Prompt for a player's name.

promptForMove :: PlayerName -> IO Cell

Prompt for move coordinates until valid input is given.

showCell :: Cell -> Game -> Char

Return a suitable display char for a given cell in a game.