TicTacToe

Documentation

data Game

A game is a mapping of points to marks and an end status.

Constructors

Game

Fields

getPlays :: Map Point Mark

getStatus :: EndStatus

data Point

A point is two coordinates.

Constructors

Point Int Int

data Mark

A mark is either X or O.

Constructors

X

O

data EndStatus

An end status describes whether a game is over, and how.

Constructors

Undecided

Winner PlayerName

type PlayerName = String

data Direction

A direction is a cardinal or ordinal point.

Constructors

N

NE

E

SE

S

SW

W

NW

Core functions

newGame :: Game

makePlay :: Mark -> Point -> Game -> Game

ticTacToe :: Mark -> Point -> Game -> Bool

Supporting functions

threeInARow :: Mark -> Point -> Direction -> Game -> Bool

inBounds :: Point -> Bool

isValidPlay :: Point -> Game -> Bool

Data.Map function aliases

markAt :: Point -> Game -> Maybe Mark

played :: Point -> Game -> Bool

notPlayed :: Point -> Game -> Bool