TicTacToe

Description

Game logic for a Tic Tac Toe game.

Documentation

data Game

A game is a mapping of cells to marks and an end status.

Constructors

Game

Fields

getMoves :: Map Cell Mark

getStatus :: EndStatus

data Cell

A cell is represented by two coordinates.

Constructors

Cell Int Int

data Mark

A mark is either an X or O.

Constructors

MarkX

MarkO

data EndStatus

An end status describes whether a game is over, and how.

Constructors

Undecided

Winner Mark

Core functions

newGame :: Game

Return an empty, undecided game.

makeMove :: Mark -> Cell -> Game -> Game

Place a mark on a game if the given cell is unoccupied.

ticTacToe :: Mark -> Cell -> Game -> Bool

Does a player win the game at a particular cell?

Supporting functions

threeInARow :: Mark -> Cell -> Direction -> Game -> Bool

Are there three of a particular mark in a row from a particular cell and direction in a game?

inBounds :: Cell -> Bool

Are some cell coordinates within the bounds of a 3x3 Tic Tac Toe board?

isValidMove :: Cell -> Game -> Bool

Is a particular cell a valid place to move in a particular game?

Data.Map aliases

markAt :: Cell -> Game -> Maybe Mark

Lookup a cell in a game, maybe returning a mark at that cell.

occupied :: Cell -> Game -> Bool

Is a particular cell in a game occupied?

unoccupied :: Cell -> Game -> Bool

Opposite of occupied.